requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Compress==1.4.0
Flask-Cors==3.0.2
gTTS==1.1.8
gTTS-token==1.1.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-aiml==0.9.0
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
